+++ xpublish_edr/geometry/area.py
@@ -55,15 +55,3 @@ def _select_area_regular_xy_grid(
 
     # Apply the mask and vectorize to a 1d collection of points
     return ds.cf.isel(X=x_sel, Y=y_sel)
-
-
-def select_by_bbox(
-    ds: xr.Dataset,
-    bbox: tuple[float, float, float, float],
-) -> xr.Dataset:
-    """
-    Return a dataset with the data within the given bbox
-
-    Assumes that the dataset is in the same CRS as the bbox
-    """
-    return ds.cf.sel(X=slice(bbox[0], bbox[2]), Y=slice(bbox[1], bbox[3]))
